+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -7,6 +7,7 @@ Release History
 * msi: use the same extension naming as portal does
 * msi: remove the useless `subscription` from the `vm/vmss create` commands output
 * `vm/vmss create`: fix a bug that the storage sku is not applied on data disks coming with an image
+* `vm format-secret`: Fix issue where `--secrets` would not accept newline separated IDs.
 
 2.0.13 (2017-08-28)
 +++++++++++++++++++